Rename package anme and namespace
