+++ tests/AuthorizationTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace NextApps\SwaggerUi\Test;
+namespace jamesRUS52\Laravel\SwaggerUi\Tests;
 
 use Illuminate\Contracts\Auth\Authenticatable;
 use Illuminate\Support\Facades\Gate;
-use NextApps\SwaggerUi\SwaggerUiServiceProvider;
+use jamesRUS52\Laravel\SwaggerUi\SwaggerUiServiceProvider;
 use Orchestra\Testbench\TestCase;
 
 class AuthorizationTest extends TestCase
@@ -26,7 +26,7 @@ class AuthorizationTest extends TestCase
      *
      * @param mixed $app
      *
-     * @return void
+     * @return string[]
      */
     protected function getPackageProviders($app)
     {
